+++ relay-server/src/managed/counted.rs
@@ -117,12 +117,6 @@ impl Counted for WithHeader<SpanV2> {
     }
 }
 
-impl Counted for SpanV2 {
-    fn quantities(&self) -> Quantities {
-        smallvec::smallvec![(DataCategory::Span, 1), (DataCategory::SpanIndexed, 1)]
-    }
-}
-
 impl Counted for Annotated<Span> {
     fn quantities(&self) -> Quantities {
         smallvec::smallvec![(DataCategory::Span, 1), (DataCategory::SpanIndexed, 1)]