+++ tests/integration/test_ourlogs.py
@@ -338,7 +338,7 @@ def test_ourlog_extraction_with_string_pii_scrubbing(
 
     relay_instance.send_envelope(project_id, envelope)
 
-    envelope = mini_sentry.captured_events.get(timeout=5)
+    envelope = mini_sentry.get_captured_event()
     item_payload = json.loads(envelope.items[0].payload.bytes.decode())
     item = item_payload["items"][0]
 
@@ -415,7 +415,7 @@ def test_ourlog_extraction_default_pii_scrubbing_attributes(
 
     relay_instance.send_envelope(project_id, envelope)
 
-    envelope = mini_sentry.captured_events.get(timeout=5)
+    envelope = mini_sentry.get_captured_event()
     item_payload = json.loads(envelope.items[0].payload.bytes.decode())
     item = item_payload["items"][0]
     attributes = item["attributes"]
@@ -457,7 +457,7 @@ def test_ourlog_default_pii_body(
 
     relay_instance.send_envelope(project_id, envelope)
 
-    envelope = mini_sentry.captured_events.get(timeout=5)
+    envelope = mini_sentry.get_captured_event()
     item_payload = json.loads(envelope.items[0].payload.bytes.decode())
     log = item_payload["items"][0]